requirements.txt
django
pygments
PyGithub
langchain
nbformat
tiktoken
cohere 
qdrant-client
chardet
faiss-cpu
streamlit